# monster/api/payment/v1/payment_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: payment/v1/payment.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPaymentCreatePayment = "/payment.v1.Payment/CreatePayment"
const OperationPaymentMarkNotificationRead = "/payment.v1.Payment/MarkNotificationRead"
const OperationPaymentPaymentCallback = "/payment.v1.Payment/PaymentCallback"

type PaymentHTTPServer interface {
	// CreatePayment 统一拉起支付
	CreatePayment(context.Context, *CreatePaymentRequest) (*CreatePaymentResponse, error)
	// MarkNotificationRead 消息已读
	MarkNotificationRead(context.Context, *MarkNotificationReadRequest) (*MarkNotificationReadResponse, error)
	// PaymentCallback 支付异步回调
	PaymentCallback(context.Context, *PaymentCallbackRequest) (*PaymentCallbackResponse, error)
}

func RegisterPaymentHTTPServer(s *http.Server, srv PaymentHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/payment/create", _Payment_CreatePayment0_HTTP_Handler(srv))
	r.POST("/api/v1/payment/callback", _Payment_PaymentCallback0_HTTP_Handler(srv))
	r.POST("/api/v1/notifications/read", _Payment_MarkNotificationRead0_HTTP_Handler(srv))
}

func _Payment_CreatePayment0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreatePaymentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentCreatePayment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreatePayment(ctx, req.(*CreatePaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreatePaymentResponse)
		return ctx.Result(200, reply)
	}
}

func _Payment_PaymentCallback0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PaymentCallbackRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentPaymentCallback)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PaymentCallback(ctx, req.(*PaymentCallbackRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PaymentCallbackResponse)
		return ctx.Result(200, reply)
	}
}

func _Payment_MarkNotificationRead0_HTTP_Handler(srv PaymentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in MarkNotificationReadRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentMarkNotificationRead)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.MarkNotificationRead(ctx, req.(*MarkNotificationReadRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MarkNotificationReadResponse)
		return ctx.Result(200, reply)
	}
}

type PaymentHTTPClient interface {
	// CreatePayment 统一拉起支付
	CreatePayment(ctx context.Context, req *CreatePaymentRequest, opts ...http.CallOption) (rsp *CreatePaymentResponse, err error)
	// MarkNotificationRead 消息已读
	MarkNotificationRead(ctx context.Context, req *MarkNotificationReadRequest, opts ...http.CallOption) (rsp *MarkNotificationReadResponse, err error)
	// PaymentCallback 支付异步回调
	PaymentCallback(ctx context.Context, req *PaymentCallbackRequest, opts ...http.CallOption) (rsp *PaymentCallbackResponse, err error)
}

type PaymentHTTPClientImpl struct {
	cc *http.Client
}

func NewPaymentHTTPClient(client *http.Client) PaymentHTTPClient {
	return &PaymentHTTPClientImpl{client}
}

// CreatePayment 统一拉起支付
func (c *PaymentHTTPClientImpl) CreatePayment(ctx context.Context, in *CreatePaymentRequest, opts ...http.CallOption) (*CreatePaymentResponse, error) {
	var out CreatePaymentResponse
	pattern := "/api/v1/payment/create"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentCreatePayment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// MarkNotificationRead 消息已读
func (c *PaymentHTTPClientImpl) MarkNotificationRead(ctx context.Context, in *MarkNotificationReadRequest, opts ...http.CallOption) (*MarkNotificationReadResponse, error) {
	var out MarkNotificationReadResponse
	pattern := "/api/v1/notifications/read"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentMarkNotificationRead))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// PaymentCallback 支付异步回调
func (c *PaymentHTTPClientImpl) PaymentCallback(ctx context.Context, in *PaymentCallbackRequest, opts ...http.CallOption) (*PaymentCallbackResponse, error) {
	var out PaymentCallbackResponse
	pattern := "/api/v1/payment/callback"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentPaymentCallback))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}