# monster/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/notifications/read:
        post:
            tags:
                - Payment
            description: 消息已读
            operationId: Payment_MarkNotificationRead
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/payment.v1.MarkNotificationReadRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/payment.v1.MarkNotificationReadResponse'
...
    /api/v1/payment/callback:
        post:
            tags:
                - Payment
            description: 支付异步回调
            operationId: Payment_PaymentCallback
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/payment.v1.PaymentCallbackRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/payment.v1.PaymentCallbackResponse'
    /api/v1/payment/create:
        post:
            tags:
                - Payment
            description: 统一拉起支付
            operationId: Payment_CreatePayment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/payment.v1.CreatePaymentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/payment.v1.CreatePaymentResponse'
...
    /api/v1/wallet/deposit/pay:
        post:
            tags:
                - User
            description: 缴纳押金
            operationId: User_PayDeposit
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.PayDepositRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.PayDepositResponse'
...
components:
    schemas:
...
        payment.v1.CreatePaymentRequest:
            type: object
            properties:
                bizType:
                    type: string
                bizNo:
                    type: string
                channel:
                    type: string
                amount:
                    type: string
                description:
                    type: string
        payment.v1.CreatePaymentResponse:
            type: object
            properties:
                paymentNo:
                    type: string
                payParams:
                    type: object
                    additionalProperties:
                        type: string
                payUrl:
                    type: string
        payment.v1.MarkNotificationReadRequest:
            type: object
            properties:
                ids:
                    type: array
                    items:
                        type: string
        payment.v1.MarkNotificationReadResponse:
            type: object
            properties:
                message:
                    type: string
        payment.v1.PaymentCallbackRequest:
            type: object
            properties:
                channel:
                    type: string
                params:
                    type: object
                    additionalProperties:
                        type: string
        payment.v1.PaymentCallbackResponse:
            type: object
            properties:
                message:
                    type: string
...
        user.v1.PayDepositRequest:
            type: object
            properties:
                amount:
                    type: string
        user.v1.PayDepositResponse:
            type: object
            properties:
                message:
                    type: string
...